tests: better ws support
cherry-picked from https://github.com/goauthentik/authentik/pull/14539
